Refactored the matplotlib-style sonification interface to support isolated per-context AudioFigure state and improve compatibility with concurrent/server-side usage.
- Moved the `sonify()` implementation into `AudioFigure` as an instance method.
- Replaced the global `_current_figure` with a Python `ContextVar`.
- Added `_get_current_figure()` utility function to retrieve (or create) the `AudioFigure` for the current execution context.
- Moved `figure_hashes` from module-level global state onto the `AudioFigure` instance.
- Removed `length` as a required constructor argument for `AudioFigure`; channel and stream setup is now initialized in `add()` on first sonification.
- Updated all user-facing convenience functions (`strauss.sonify()`, `strauss.save()`, etc.) to use `_get_current_figure()` internally.
