+++ src/strauss/audio_figure.py
@@ -1,3 +1,29 @@
+# relative imports of submodules
+from . import generator as generator_module # Renamed to avoid conflict with 'generator' variable
+from . import score as score_module # Renamed to avoid conflict with 'score' variable
+from . import sonification as sonification_module # Renamed to avoid conflict with 'sonification' variable
+from . import sources as sources_module # Renamed to avoid conflict with 'sources' variable
+
+# relative imports of submodules
+from . import channels
+from . import filters
+from . import generator
+from . import notes
+from . import score
+from . import sonification
+from . import sources
+from . import stream
+from . import presets
+from . import styles
+from . import assets
+
+# Import core classes for direct use in the sonify function
+from .score import Score
+from .sources import Events, Objects, set_limits
+from .generator import Synthesizer, Sampler, Spectralizer
+from .sonification import Sonification
+from .utilities import nested_dict_reassign, merge_events, rescale_values
+
 import numpy as np
 from . import channels
 from .stream import Stream
@@ -15,26 +41,96 @@ try:
         from tqdm import tqdm
 except ModuleNotFoundError:
     tqdm = list
+    
+import yaml
+import glob
+import warnings
+from pathlib import Path
+import hashlib
+import json
 
+p = Path(__file__)
+thisdir = p.parent
 
 INTMAX32 = (pow(2, 31)-1)
+
+_kw_defaults = {
+    'channels': 'stereo',
+    'duration': 10,
+    'is_mapped': ['pitch', 'time_evo'],
+    # Style File
+    'style' : None,
+    'caption': None,
+    'name': None,
+    'level': 1,
+}
+
+_exclude_keys = ['name', 'level']
+
+def fill_from_kwargs(input_kwargs):
+        """ function to store provided keyword arguments against defaults
+        """
+        sonpars = input_kwargs.copy()
+        is_default = {}
+        for k in _kw_defaults.keys():
+            if k not in sonpars:
+                sonpars[k] = _kw_defaults[k]
+                is_default[k] = True
+            else:
+                is_default[k] = False
+        return sonpars, is_default
+    
+def _get_style_path(name="default"):
+        path = Path(name)
+        if (len(path.parts) > 1) or (path.suffix != ''):
+            # if open user directly
+            return Path(name)
+        else:
+            # else load built-in preset of that name
+            return Path(f"{thisdir}", "styles", f"{name}.yml")
+        
+def load_style(name="default"):
+    filename = _get_style_path(name)
+    return read_yaml(filename)
+
+def get_style(name, print_style=False):
+    filename = _get_style_path(name)
+    with filename.open(mode='r') as fdata:
+        # strip unnecessary whitespace
+        yaml_string = fdata.read().rstrip().lstrip()
+    if print_style:
+        print(yaml_string)
+    return yaml_string
+        
+def read_yaml(filename):
+    with filename.open(mode='r') as fdata:
+        try:
+            yamldict = yaml.safe_load(fdata)
+        except yaml.YAMLError as err:
+            print(err)
+    return yamldict
+
+
+def arg_hash(arg):
+    if isinstance(arg, np.ndarray):
+        return hashlib.md5(arg.tobytes()).hexdigest()
+    return str(arg)
+
 
 class AudioFigure:
     """
     A figure-like wrapper to manage, mix, and render multiple STRAUSS sonifications.
     """
-    def __init__(self, length, samprate=48000, system="stereo"):
-        self.length = length             # Master duration in seconds
+    def __init__(self, samprate=48000, system="stereo"):
+        self.length = None             # Master duration in seconds
         self.samprate = samprate         # Master sampling rate
         self.system = system             # Master channel format (mono/stereo/etc.)
-
-        # We set up the master channels for the audio figure
-        self.channels = channels.audio_channels(setup=system)
-
-        # with the individual channel arrays
-        self.channel_arrays = {}
-        for c in range(self.channels.Nmics):
-            self.channel_arrays[str(c)] = Stream(self.length, self.samprate)
+        
+        self.figure_hashes = {}
+        self.style_cycle_index = 0
+        self.style_cycle = {"theremin",
+                            "windy",
+                            "clicker"}
 
         # Store sonifications in a dict for easy access and modification
         self.sonifications = {}
@@ -50,13 +146,173 @@ class AudioFigure:
 
         # has this figure been rendered?
         self.is_rendered = False
+    
         
+    def sonify(self, *args, **kwargs):
+        """
+        Generate a sonification in a matplotlib-like interface and add it to the AudioFigure.
+
+        Args:
+            data: The input data for sonification.
+            style (dict, optional): A dictionary defining the sonification style,
+                                    including parameters for score, sources, and generator.
+            **kwargs: Additional parameters to override or supplement the style.
+
+        Returns:
+            The generated Sonification object.
+        """
+        
+        sonpars, is_default = fill_from_kwargs(kwargs)            
+
+        # We first analyse if this is a repeat sonification...
+        # Initially filter relevent keys
+        kwargs_to_hash = {k: v for k, v in sonpars.items() if k not in _exclude_keys}
+        args_to_hash = [arg_hash(arg) for arg in args]
+        
+        # Put it in a dictionary, and get a unique hash
+        param_dict = {
+            'args': args_to_hash, 
+            'kwargs': kwargs_to_hash  # Use the filtered dict here
+        }
+        param_str = json.dumps(param_dict, sort_keys=True)
+        current_hash = hashlib.md5(param_str.encode('utf-8')).hexdigest()
+
+        # now check against existing hashes and rename sonification
+        # or re-set volume level if necessary
+        if current_hash in self.figure_hashes.keys():
+            name = self.figure_hashes[current_hash]
+            mssg = 'Matching sonification exists, '
+            actions = []
+            if ('name' in sonpars) and sonpars['name'] and sonpars['name'] != name:
+                actions.append('renaming')
+                self.rename(old=name, new=sonpars['name'])
+                name = sonpars['name']
+            if (sonpars['level'] != self.levels[name]):
+                actions.append(f"re-setting level to {sonpars['level']}")
+                self.set_level(name, sonpars['level'])
+            if not actions:
+                actions.append('skipping')
+            print(mssg + ' and '.join(actions) + '...')
+            
+            # replace the named hash
+            self.figure_hashes[current_hash] = name
+            return self.sonifications[name]
+                
+        if not sonpars['style']:
+            sonpars['style'] = 'default'
+        style = styles.Style(**load_style(sonpars['style']))
+        
+        if len(args) == 0:
+            raise Exception("No data to sonify!")
+        elif len(args) == 1:
+            args = [np.arange(len(args[0])), args[0]]
+            tlims = (0,args[0][-1])
+
+        if (style.sources.lower() == 'events') and style.max_notes_per_sec:
+            if style.map[0].output == 'time':
+                tlims = style.map[0].input_range
+            tlims = set_limits(tlims, args[0], warn=False)
+            time = rescale_values(args[0], tlims, (0,1))
+            
+            # lets now thin the data according to max events per second if using events
+            args = merge_events(sonpars['duration'], style.max_notes_per_sec, time, args)
+                
+        nmap = min(len(args), len(style.map))
+        
+        to_map = []
+        in_lims = {}
+        out_lims = {}
+        for i in range(nmap):
+            to_map.append(style.map[i].output)
+            if style.map[i].input_range:
+                in_lims[to_map[-1]] = style.map[i].input_range
+            if style.map[i].output_range:
+                out_lims[to_map[-1]] = style.map[i].output_range
+        map_data = dict(zip(to_map, args[:nmap]))
+        
+        if 'pitch' not in to_map:
+            to_map.append('pitch')
+            nnote = len(style.notes)
+            for k in map_data.keys():
+                map_data[k] = [map_data[k]]*nnote
+            map_data['pitch'] = list(range(nnote))
+            to_map.append('pitch')
+        
+        # we now iterate through style fixed values
+        if len(args) > nmap: 
+            for i in range(nmap, len(args)):
+                if style.map[i].fixed:
+                    if style.map[i].input_range:
+                        in_lims[to_map[-1]] = style.map[i].input_range
+                    if style.map[i].output_range:
+                        out_lims[to_map[-1]] = style.map[i].output_range
+
+                    fix_array =  len(map_data[to_map[0]])*[style.map[i].fixed]
+                    to_map.append(style.map[i].output)
+                    map_data[style.map[i].output] = fix_array
+                    
+        # and finally overwrite with any kwarg fixed values:
+        for k in sonpars.keys():
+            ksplit = k.split('fix_')
+            if len(ksplit) > 1:
+                prop = ksplit[1]
+                if prop in to_map:
+                    print(f'Overwriting {prop} with fixed value...')
+                map_data[prop] = [sonpars[k]]*len(map_data[to_map[0]])
+                if prop in ['azimuth', 'polar']:
+                    if prop == 'polar':
+                        in_lims[prop] = (0,180)
+                    if prop == 'azimuth':
+                        in_lims[prop] = (0,360)
+                else:
+                    in_lims[prop] = (0,1)
+                to_map.append(prop)
+
+        snotes = style.notes
+        if not isinstance(style.notes, str):
+            snotes = [snotes]
+        _score = Score(snotes, length=sonpars['duration'])
+        _sources = getattr(sources, style.sources.capitalize())(to_map)
+        _sources.fromdict(map_data)
+        _sources.apply_mapping_functions(map_lims=in_lims, param_lims=out_lims)
+
+        gentype = style.generator.type
+        
+        if gentype == 'sampler':
+            s = style.generator.sample
+            if '/' in s or '\\' in s or '.' in s:
+                # this is probably intended to be a file path
+                asset = s
+            else:
+                # if not, assume intention is in-built name
+                asset = assets.get_asset_path(s.lower())
+            _generator = getattr(generator, "Sampler")(asset)
+        else:
+            _generator = getattr(generator, style.generator.type.capitalize())()
+        _generator.load_preset(style.generator.preset)
+        if style.generator.mods:
+            _generator.modify_preset(style.generator.mods)
+        _sonification = Sonification(
+            score=_score,
+            sources=_sources,
+            generator=_generator,
+            audio_setup=sonpars['channels'],
+            caption=sonpars["caption"],
+            samprate=_generator.samprate, # Use generator's samprate for consistency
+        )
+            
+        name = self.add(_sonification, name=sonpars["name"], level=sonpars["level"])
+        self.figure_hashes[current_hash] = name
+        
+        return _sonification
+    
+    
     def list_sonifications(self):
         for i, k in enumerate(self.sonifications.keys()):
             print(f"\t{i+1}.\t {k}")
 
     def rename(self, old, new):
-        dicts = [self.sonifications, self.levels, self.styles, self.style_hashes]
+        dicts = [self.sonifications, self.levels, self.styles, self.figure_hashes]
         for d in dicts:
             d[new] = d[old]
             del d[old]
@@ -74,8 +330,20 @@ class AudioFigure:
         # Enforce global timebase on the child
         soni.samprate = self.samprate
         soni.system = self.system
-        # (Assuming the length is set on the generator, score, or sonification itself in your architecture)
-        # soni.generator.length = self.length 
+        
+        # Set up channels and channel_arrays on first sonification
+        if self.length is None:
+            self.length = soni.score.length
+
+            self.channels = channels.audio_channels(setup=self.system)
+
+            self.channel_arrays = {}
+
+            for c in range(self.channels.Nmics):
+                self.channel_arrays[str(c)] = Stream(
+                    self.length,
+                    self.samprate
+                )
 
         if self.sonifications:
             # preliminary check if sonification is consistent with existing timebase
@@ -139,7 +407,7 @@ class AudioFigure:
             raise KeyError(f"Sonification '{name}' not found in AudioFigure.")
 
     def _parse_level(self, level):
-        """
+        """ 
         Parses a mixing level to a linear amplitude fraction.
         
         Supports:
@@ -211,8 +479,8 @@ class AudioFigure:
 
         # 6. Apply clipping protection
         # self._apply_normalization(normalize)
-        vmax = abs(self.master_audio).max()
-        self.master_audio *= (pow(2, 31)-1)/vmax
+        self.vmax = abs(self.master_audio).max()
+        self.master_audio *= (pow(2, 31)-1)/self.vmax
         self.master_audio = self.master_audio.astype('int32')
         self.is_rendered = True
         
@@ -230,8 +498,8 @@ class AudioFigure:
         if has_ticks:
             # add the ticks
             for c in range(outfmt.shape[0]):
-                outfmt[c] += self.tick_channels['0'].values*self.tick_vol / vmax
-        display(ipd.Audio(outfmt.T,rate=self.samprate, autoplay=False))
+                outfmt[c] += self.tick_channels['0'].values*self.tick_vol / self.vmax
+        ipd.display(ipd.Audio(outfmt.T,rate=self.samprate, autoplay=False))
 
     def save(self, fname):
         # combine and write out file. first check extension
